Better alias handling (kinda)
Former-commit-id: d999d9b9e1803c2fb93b88891108496370c9d33c

+++ .config/zsh/aliases.zsh
@@ -51,6 +51,8 @@ alias yta="yt -x -f bestaudio/best"
 alias ytdl="youtube-dl"
 
 #replace root-only commands with sudo [command]
-for command in mount umount sv visudo pacman updatedb su shutdown poweroff reboot ; do
-	alias $command="sudo $command"
-done; unset command
+if [ -f /bin/sudo ];
+	for command in mount umount sv visudo pacman updatedb su shutdown poweroff reboot ; do
+		alias $command="sudo $command"
+	done; unset command
+fi
